requirements.txt
requests
absl-py